requirements.txt
fastapi~=0.95.2
natsort~=8.3.1
starlette~=0.27.0
uvicorn
openpyxl
pandas~=1.4.2
requests
jinja2
pyjwt
python-decouple
matplotlib~=3.7.1
python-dotenv~=0.21.1
gspread~=5.8.0
json2html~=1.3.0
pymssql~=2.2.7
python-multipart
pydantic~=1.10.4
aiohttp~=3.7.4.post0
starlette-discord~=0.2.1
itsdangerous~=2.1.2
fastapi-sessions~=0.3.2